# openapi.yaml
components:
  schemas:
...
    EntityCreateRequest:
      properties:
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/EntityMeta'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - meta
      - content
      title: EntityCreateRequest
      type: object
    EntityMeta:
      properties:
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        namespace:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace
        tags:
          items:
            type: string
          title: Tags
          type: array
        when_to_use:
          anyOf:
          - type: string
          - type: 'null'
          title: When To Use
      required:
      - name
      title: EntityMeta
      type: object
...
    EntityUpdateRequest:
      properties:
        change_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Change Summary
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          anyOf:
          - $ref: '#/components/schemas/EntityMeta'
          - type: 'null'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - content
      title: EntityUpdateRequest
      type: object
    EvolutionMeta:
      properties:
        behavioral_descriptors:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Behavioral Descriptors
        fitness:
          anyOf:
          - type: number
          - type: 'null'
          title: Fitness
        is_valid:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Valid
        metrics:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metrics
        mutation_kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Mutation Kind
        prompt_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt Ref
      title: EvolutionMeta
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    MemoryCardResponse:
      properties:
        channel:
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        entity_id:
          title: Entity Id
          type: string
        entity_type:
          const: memory_card
          default: memory_card
          title: Entity Type
          type: string
        etag:
          title: Etag
          type: string
        meta:
          additionalProperties: true
          title: Meta
          type: object
        version_id:
          title: Version Id
          type: string
      required:
      - entity_id
      - version_id
      - channel
      - etag
      - meta
      - content
      title: MemoryCardResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: 'Persistent memory for CARL artifacts: steps, chains, agents, memory
    cards'
  title: GigaEvo Memory Module
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/events/stream:
    get:
      description: Server-Sent Events stream for entity change notifications.
      operationId: event_stream_v1_events_stream_get
      parameters:
      - in: query
        name: entity_type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Type
      - in: query
        name: entity_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Id
      - in: query
        name: namespace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Event Stream
      tags:
      - events
  /v1/maintenance/clear-all:
    post:
      description: "Soft-delete all entities, optionally filtered by type.\n\nArgs:\n\
        \    entity_type: Optional entity type to clear (step, chain, agent, memory_card).\n\
        \                 If not provided, clears all entity types.\n\nReturns:\n\
        \    Dictionary with counts of deleted entities per type."
      operationId: clear_all_entities_v1_maintenance_clear_all_post
      parameters:
      - in: query
        name: entity_type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Type
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Clear All Entities
      tags:
      - entities (deprecated)
  /v1/memory-cards:
    get:
      description: List all memory cards with pagination.
      operationId: list_memory_cards_v1_memory_cards_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MemoryCardResponse'
                title: Response List Memory Cards V1 Memory Cards Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Memory Cards
      tags:
      - memory_cards
    post:
      description: Create a new memory card entity with its first version.
      operationId: create_memory_card_v1_memory_cards_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryCardResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Memory Card
      tags:
      - memory_cards
  /v1/memory-cards/{memory_card_id}:
    delete:
      description: Soft-delete a memory card.
      operationId: delete_memory_card_v1_memory_cards__memory_card_id__delete
      parameters:
      - in: path
        name: memory_card_id
        required: true
        schema:
          format: uuid
          title: Memory Card Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Memory Card
      tags:
      - memory_cards
    get:
      description: Get a memory card by ID, resolving the channel to a specific version.
      operationId: get_memory_card_v1_memory_cards__memory_card_id__get
      parameters:
      - in: path
        name: memory_card_id
        required: true
        schema:
          format: uuid
          title: Memory Card Id
          type: string
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      - in: header
        name: if-none-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-None-Match
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryCardResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Memory Card
      tags:
      - memory_cards
    put:
      description: Update a memory card by creating a new immutable version.
      operationId: update_memory_card_v1_memory_cards__memory_card_id__put
      parameters:
      - in: path
        name: memory_card_id
        required: true
        schema:
          format: uuid
          title: Memory Card Id
          type: string
      - in: header
        name: if-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryCardResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Memory Card
      tags:
      - memory_cards
...
  /v1/webhooks/{webhook_id}:
    delete:
      description: Delete a webhook (not yet implemented).
      operationId: delete_webhook_v1_webhooks__webhook_id__delete
      parameters:
      - in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
      responses:
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '501':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Delete Webhook
      tags:
      - events
...